refactor page templates
