+++ src/page-templates/content-type-page/content-type-data.ts
