+++ src/page-templates/content-type-page/index.ts
