+++ src/page-templates/content-type-page/styles.ts
