+++ src/page-templates/content-type-page/types.ts
